Refactor: Eggpop type
