+++ src/app/(pages)/(club)/club/_api/clubSubmission.ts
@@ -1,5 +1,5 @@
 import browserClient from "@/utils/supabase/client";
-import { EggPopForm } from "@/types/안끝난거/eggpop.types";
+import { EggPopForm } from "@/types/eggpop.types";
 import { EggClubForm } from "@/types/eggclub.types";
 
 export const submitOneTimeClubData = async (finalFormData: EggPopForm) => {
